requirements.txt
fastapi
uvicorn
openai
sqlalchemy
pydantic
qdrant-client
python-dotenv